backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attackcti==0.5.4
attrs==25.3.0
Automat==24.8.1
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blis==1.2.0
Brotli==1.1.0
build==1.2.2.post1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
contractions==0.1.73
crochet==2.1.1
cryptography==44.0.2
cssselect==1.3.0
cvss==3.4
cymem==2.0.11
decorator==5.2.1
defusedxml==0.7.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
imbalanced-learn==0.13.0
incremental==24.7.2
ipython==8.12.3
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.9.2
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
murmurhash==1.0.12
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pickleshare==0.7.5
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.3.7
preshed==3.0.9
prompt_toolkit==3.0.50
Protego==0.4.0
protobuf==5.29.4
pure_eval==0.2.3
pyahocorasick==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.7.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-tor==1.4
rich==13.9.4
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.15.2
Scrapy==2.12.0
service-identity==24.2.0
setuptools==76.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
stem==1.8.2
stix2==3.0.1
stix2-patterns==2.0.0
sympy==1.13.1
taxii2-client==2.3.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
textsearch==0.0.24
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.21.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
Twisted==24.11.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vaderSentiment==3.3.2
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
yarg==0.1.9
zope.interface==7.2
